Explicit MinimizeTest
