+++ src/functionalTest/kotlin/com/github/jengelman/gradle/plugins/shadow/BasePluginTest.kt
@@ -216,79 +216,6 @@ abstract class BasePluginTest {
     )
   }
 
-  fun writeApiLibAndImplModules() {
-    settingsScriptPath.appendText(
-      """
-        include 'api', 'lib', 'impl'
-      """.trimIndent() + System.lineSeparator(),
-    )
-    projectScriptPath.writeText("")
-
-    path("lib/src/main/java/lib/LibEntity.java").writeText(
-      """
-        package lib;
-        public interface LibEntity {}
-      """.trimIndent(),
-    )
-    path("lib/src/main/java/lib/UnusedLibEntity.java").writeText(
-      """
-        package lib;
-        public class UnusedLibEntity implements LibEntity {}
-      """.trimIndent(),
-    )
-    path("lib/build.gradle").writeText(
-      """
-        plugins {
-          id 'java'
-        }
-      """.trimIndent() + System.lineSeparator(),
-    )
-
-    path("api/src/main/java/api/Entity.java").writeText(
-      """
-        package api;
-        public interface Entity {}
-      """.trimIndent(),
-    )
-    path("api/src/main/java/api/UnusedEntity.java").writeText(
-      """
-        package api;
-        import lib.LibEntity;
-        public class UnusedEntity implements LibEntity {}
-      """.trimIndent(),
-    )
-    path("api/build.gradle").writeText(
-      """
-        plugins {
-          id 'java'
-        }
-        dependencies {
-          implementation 'junit:junit:3.8.2'
-          implementation project(':lib')
-        }
-      """.trimIndent() + System.lineSeparator(),
-    )
-
-    path("impl/src/main/java/impl/SimpleEntity.java").writeText(
-      """
-        package impl;
-        import api.Entity;
-        public class SimpleEntity implements Entity {}
-      """.trimIndent(),
-    )
-    path("impl/build.gradle").writeText(
-      """
-        ${getDefaultProjectBuildScript("java-library")}
-        dependencies {
-          api project(':api')
-        }
-        $shadowJar {
-          minimize()
-        }
-      """.trimIndent() + System.lineSeparator(),
-    )
-  }
-
   fun writeGradlePluginModule(legacy: Boolean) {
     val pluginId = "my.plugin"
     val pluginClass = "my.plugin.MyPlugin"
