+++ src/functionalTest/kotlin/com/github/jengelman/gradle/plugins/shadow/JavaPluginTest.kt
@@ -150,202 +150,6 @@ class JavaPluginTest : BasePluginTest() {
     }
   }
 
-  /**
-   * 'Server' depends on 'Client'. 'junit' is independent.
-   * The minimize shall remove 'junit'.
-   */
-  @Test
-  fun minimizeByKeepingOnlyTransitiveDependencies() {
-    writeClientAndServerModules(
-      serverShadowBlock = """
-        minimize()
-      """.trimIndent(),
-    )
-    path("server/src/main/java/server/Server.java").writeText(
-      """
-        package server;
-        import client.Client;
-        public class Server {
-          // This is to make sure that 'Client' is not removed.
-          private final String client = Client.class.getName();
-        }
-      """.trimIndent(),
-    )
-
-    run(serverShadowJarTask)
-
-    assertThat(outputServerShadowJar).useAll {
-      containsEntries(
-        "client/Client.class",
-        "server/Server.class",
-      )
-      doesNotContainEntries(
-        "junit/framework/Test.class",
-      )
-    }
-  }
-
-  /**
-   * 'Client', 'Server' and 'junit' are independent.
-   * 'junit' is excluded from the minimize step.
-   * The minimize step shall remove 'Client' but not 'junit'.
-   */
-  @Test
-  fun excludeDependencyFromMinimize() {
-    writeClientAndServerModules(
-      serverShadowBlock = """
-        minimize {
-          exclude(dependency('junit:junit:.*'))
-        }
-      """.trimIndent(),
-    )
-
-    run(serverShadowJarTask)
-
-    assertThat(outputServerShadowJar).useAll {
-      containsEntries(
-        "server/Server.class",
-        "junit/framework/Test.class",
-      )
-      doesNotContainEntries(
-        "client/Client.class",
-      )
-    }
-  }
-
-  /**
-   * 'Client', 'Server' and 'junit' are independent.
-   * Unused classes of 'client' and theirs dependencies shouldn't be removed.
-   */
-  @Test
-  fun excludeProjectFromMinimize() {
-    writeClientAndServerModules(
-      serverShadowBlock = """
-        minimize {
-          exclude(project(':client'))
-        }
-      """.trimIndent(),
-    )
-
-    run(serverShadowJarTask)
-
-    assertThat(outputServerShadowJar).useAll {
-      containsEntries(
-        "client/Client.class",
-        "server/Server.class",
-      )
-    }
-  }
-
-  /**
-   * 'Client', 'Server' and 'junit' are independent.
-   * Unused classes of 'client' and theirs dependencies shouldn't be removed.
-   */
-  @Test
-  fun excludeProjectFromMinimizeShallNotExcludeTransitiveDependenciesThatAreUsedInSubproject() {
-    writeClientAndServerModules(
-      serverShadowBlock = """
-        minimize {
-          exclude(project(':client'))
-        }
-      """.trimIndent(),
-    )
-    path("client/src/main/java/client/Client.java").writeText(
-      """
-        package client;
-        import junit.framework.TestCase;
-        public class Client extends TestCase {
-          public static void main(String[] args) {}
-        }
-      """.trimIndent(),
-    )
-
-    run(serverShadowJarTask)
-
-    assertThat(outputServerShadowJar).useAll {
-      containsEntries(
-        "client/Client.class",
-        "server/Server.class",
-        "junit/framework/TestCase.class",
-      )
-    }
-
-    path("client/src/main/java/client/Client.java").writeText(
-      """
-        package client;
-        public class Client {}
-      """.trimIndent(),
-    )
-    run(serverShadowJarTask)
-
-    assertThat(outputServerShadowJar).useAll {
-      containsEntries(
-        "client/Client.class",
-        "server/Server.class",
-        "junit/framework/TestCase.class",
-      )
-    }
-  }
-
-  /**
-   * 'api' used as api for 'impl', and depended on 'lib'. 'junit' is independent.
-   * The minimize step shall remove 'junit', but not 'api'.
-   * Unused classes of 'api' and theirs dependencies also shouldn't be removed.
-   */
-  @Test
-  fun useMinimizeWithDependenciesWithApiScope() {
-    writeApiLibAndImplModules()
-
-    run(":impl:$SHADOW_JAR_TASK_NAME")
-
-    assertThat(jarPath("impl/build/libs/impl-all.jar")).useAll {
-      containsEntries(
-        "impl/SimpleEntity.class",
-        "api/Entity.class",
-        "api/UnusedEntity.class",
-        "lib/LibEntity.class",
-      )
-      doesNotContainEntries(
-        "junit/framework/Test.class",
-        "lib/UnusedLibEntity.class",
-      )
-    }
-  }
-
-  /**
-   * 'api' used as api for 'impl', and 'lib' used as api for 'api'.
-   * Unused classes of 'api' and 'lib' shouldn't be removed.
-   */
-  @Test
-  fun useMinimizeWithTransitiveDependenciesWithApiScope() {
-    writeApiLibAndImplModules()
-    path("api/build.gradle").writeText(
-      """
-        plugins {
-          id 'java-library'
-        }
-        dependencies {
-          api project(':lib')
-        }
-      """.trimIndent(),
-    )
-
-    run(":impl:$SHADOW_JAR_TASK_NAME")
-
-    assertThat(jarPath("impl/build/libs/impl-all.jar")).useAll {
-      containsEntries(
-        "impl/SimpleEntity.class",
-        "api/Entity.class",
-        "api/UnusedEntity.class",
-        "lib/LibEntity.class",
-        "lib/UnusedLibEntity.class",
-      )
-      doesNotContainEntries(
-        "junit/framework/Test.class",
-      )
-    }
-  }
-
   @Test
   fun dependOnProjectShadowJar() {
     writeShadowedClientAndServerModules()
